+++ build.gradle
@@ -1,98 +1,137 @@
+import org.gradle.api.tasks.testing.logging.TestExceptionFormat
+
 buildscript {
-	repositories {
-		mavenCentral()
-	}
+    repositories {
+        mavenCentral()
+    }
 }
 
 plugins {
-	id 'java'
-	id 'idea'
-	id 'org.springframework.boot' version '2.2.2.RELEASE'
-	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
-	id 'com.google.cloud.tools.jib' version '1.8.0'
-	id 'org.openapi.generator' version '4.2.3'
-	id 'com.diffplug.gradle.spotless' version '3.27.2'
+    id 'java'
+    id 'idea'
+    id 'com.google.cloud.tools.jib' version '2.5.0'
+    id 'com.diffplug.gradle.spotless' version '3.27.2'
+    id 'de.undercouch.download' version '4.0.0'
+    id 'io.spring.dependency-management' version '1.0.9.RELEASE'
+    id 'org.springframework.boot' version '2.3.1.RELEASE'
+    id 'org.hidetake.swagger.generator' version '2.18.2'
 }
 
-group = 'bio.terra.TEMPLATE'
-version = '0.0.1-SNAPSHOT'
+group = 'bio.terra'
+version = System.getenv('SERVICE_VERSION') != null ? System.getenv('SERVICE_VERSION') : '0.0.1-SNAPSHOT'
 sourceCompatibility = '1.8'
 
 repositories {
-	mavenCentral()
+    mavenCentral()
 }
 
 dependencies {
-	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jdbc'
-	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web'
-	implementation group: 'org.postgresql', name: 'postgresql', version: '42.1.4'
-	implementation group: 'org.apache.commons', name: 'commons-dbcp2', version: '2.7.0'
-	implementation group: 'org.apache.commons', name: 'commons-pool2', version: '2.8.0'
-	implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
-	implementation group: 'org.liquibase' , name: 'liquibase-core', version: '3.8.6'
-	implementation group: 'org.webjars', name: 'swagger-ui', version: '3.24.0'
+    implementation group: 'javax.validation', name: 'validation-api', version: '2.0.1.Final'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jdbc'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web'
+    implementation group: 'org.webjars', name: 'swagger-ui', version: '3.25.4' // if updating, change path in ApiResourceConfig.java
+    implementation group: 'org.apache.commons', name: 'commons-dbcp2', version: '2.7.0'
+    implementation group: 'org.apache.commons', name: 'commons-pool2', version: '2.8.0'
+    implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.10'
+    implementation group: 'org.postgresql', name: 'postgresql', version: '42.2.12'
+    implementation group: 'org.liquibase' , name: 'liquibase-core', version: '3.10.0'
+    implementation 'org.springframework.boot:spring-boot-starter'
+    implementation 'org.springframework:spring-jdbc:5.1.9.RELEASE'
 
-	// -- OpenAPI CodeGen dependencies --
-	// TODO: this version of swagger-annotations is old, but the code gen is still relying on it
-	//  There is an open bug tracking the fix: https://github.com/OpenAPITools/openapi-generator/issues/4245
-	implementation group: 'io.swagger', name: 'swagger-annotations', version: '1.6.0'
-	implementation group: 'org.openapitools', name: 'jackson-databind-nullable', version: '0.2.1'
-	// -- --
+    implementation 'com.google.guava:guava:29.0-jre'
 
-	testImplementation('org.springframework.boot:spring-boot-starter-test') {
-		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
-	}
+    // -- Swagger CodeGen dependencies --
+    swaggerCodegen 'io.swagger.codegen.v3:swagger-codegen-cli:3.0.20'
+    implementation group: 'io.swagger.core.v3', name: 'swagger-annotations', version: '2.1.4'
+    // TODO: this version of swagger-annotations is old, but the code gen is still relying on it
+    // There is an open bug tracking the fix: https://github.com/OpenAPITools/openapi-generator/issues/4245
+    implementation 'io.swagger:swagger-annotations:1.6.0'
+
+    testImplementation('org.springframework.boot:spring-boot-starter-test') {
+        exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+    }
 }
 
 // Exclude the Spring logger, so everything will use SLF4J
 configurations.all {
-	exclude group: "commons-logging", module: "commons-logging"
+    exclude group: "commons-logging", module: "commons-logging"
 }
+
+java {
+    withJavadocJar()
+}
+
+// Linter
+spotless {
+    java {
+        googleJavaFormat()
+    }
+}
+
 
 // OpenAPI Server Generation
-// - specify the source and target
-// - define the generation
-// - define the validation
 // - define dependencies to include the generated code
-
-def openapiSourceFile = "${projectDir}/src/main/resources/api/service_openapi.yaml"
+def openapiSourceDir = "${projectDir}/src/main/resources/api"
 def openapiTargetDir = "${buildDir}/generated"
-
-openApiGenerate {
-	generatorName = "spring"
-	inputSpec = "${openapiSourceFile}".toString()
-	outputDir = "${openapiTargetDir}".toString()
-	packageName = "${group}"
-	apiPackage = "${group}.generated.controller"
-	modelPackage = "${group}.generated.model"
-	configOptions = [
-	        interfaceOnly : "true",
-			useTags : "true",
-			library : "spring-boot",
-			dateLibrary : "java8",
-			generateApiTests : "false",
-			generateApiDocumentation : "false",
-			generateModelTests : "false",
-			generateModelDocumentation : "false",
-	]
+swaggerSources {
+    server {
+        inputFile = file("${openapiSourceDir}/service_openapi.yaml")
+        code {
+            language = "spring"
+            outputDir = file("${openapiTargetDir}")
+            components = ["models", "apis"]
+            rawOptions = [
+                    "--api-package", "${this.group}.rbs.generated.controller",
+                    "--model-package", "${this.group}.rbs.generated.model",
+                    "--library", "spring-boot",
+                    "-D", "interfaceOnly=true," +
+                            "useTags=true," +
+                            "dateLibrary=java8"
+            ]
+        }
+    }
 }
 
-openApiValidate {
-	inputSpec = "${openapiSourceFile}".toString()
-}
+// Note: Spotless deletes the classes OpenApi generates.
+// The order needs to be spotlessApply -> openApiGenerate -> build/test/run
+swaggerSources.server.code.dependsOn tasks.spotlessApply
+compileJava.dependsOn swaggerSources.server.code, spotlessApply
 
-compileJava.dependsOn tasks.openApiGenerate
 sourceSets.main.java.srcDir "${openapiTargetDir}/src/main/java"
-ideaModule.dependsOn tasks.openApiGenerate
-
+ideaModule.dependsOn swaggerSources.server.code
 // end of OpenAPI Server Generation
 
-test {
-	useJUnitPlatform()
+tasks.withType(Test) {
+    testLogging {
+        events = ["passed", "failed", "skipped"]
+        // Causes the correct line to be reported on an exception.
+        exceptionFormat = TestExceptionFormat.FULL
+        reports {
+            // Write XML file (used by CircleCI, Jenkins, etc) to build/test-results/*
+            junitXml.enabled = true
+            // Write human-readable test report to build/reports/tests/*
+            html.enabled = true
+        }
+    }
 }
 
-spotless {
-	java {
-		googleJavaFormat()
-	}
+test {
+    useJUnitPlatform {
+        includeTags 'unit', 'integration'
+    }
+}
+
+task unitTest(type: Test) {
+    useJUnitPlatform {
+        includeTags 'unit'
+    }
+}
+
+task integrationTest(type: Test) {
+    useJUnitPlatform {
+        includeTags "integration"
+    }
+    // Force tests to always be re-run, since integration tests involve communicating with external
+    // resources.
+    outputs.upToDateWhen { false }
 }
