+++ src/main/java/bio/terra/rbs/app/configuration/BeanNames.java
@@ -1,0 +1,9 @@
+package bio.terra.rbs.app.configuration;
+
+/** Bean name constants */
+public class BeanNames {
+  private BeanNames() {}
+
+  public static final String JDBC_TEMPLATE = "jdbcTemplate";
+  public static final String OBJECT_MAPPER = "objectMapper";
+}
