--- src/main/resources/application.properties
@@ -1,7 +1,0 @@
-server.port=8080
-template.dbUri=jdbc:postgresql://${DATABASE_HOSTNAME:127.0.0.1}:5432/${DATABASE_NAME}
-template.dbUsername=${DATABASE_USER}
-template.dbPassword=${DATABASE_USER_PASSWORD}
-template.dbInitializeOnStart=false
-template.dbInitialize=false
-template.dbUpgradeOnStart=true
