+++ src/test/java/bio/terra/rbs/common/BaseUnitTest.java
@@ -1,0 +1,16 @@
+package bio.terra.rbs.common;
+
+import bio.terra.rbs.app.Main;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+@Tag("unit")
+@ActiveProfiles({"test", "unit"})
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = Main.class)
+@SpringBootTest
+public class BaseUnitTest {}
