+++ src/test/resources/application-test.yml
@@ -1,0 +1,7 @@
+rbs:
+  db:
+    password: dbpwd
+    recreate-db-on-start: true
+    update-db-on-start: true
+    uri: jdbc:postgresql://127.0.0.1:5432/testdb
+    username: dbuser
