+++ src/test/resources/application-unit.yml
@@ -1,0 +1,1 @@
+# empty
