--- src/test/resources/application.properties
@@ -1,6 +1,0 @@
-server.port=8080
-template.dbUri=jdbc:postgresql://127.0.0.1:5432/template
-template.dbUsername=templatemanager
-template.dbPassword=TEMPLATEpassword
-template.dbInitialize=true
-template.dbUpgradeOnStart=true
